Implemented planning functionality.
The user selects which projects they will work on for the day on the project table.

Once any number of projects are selected, the instruction box at the top will show that the planning for the day is complete.

The selected projects for the day are tracked in the session[:today][:projects] set, and updated via AJAX when the user selects or deselects rows from the table.

This is the big commit for #81